Fix class cast exception by parsing MTKHEREFLIP as BigInt instead of Int

+++ digiroad2-oracle/src/main/scala/fi/liikennevirasto/digiroad2/util/AssetDataImporter.scala
@@ -1069,7 +1069,7 @@ class AssetDataImporter {
   private def isHereFlipped(roadLink: VVHRoadlink) = {
     val NotFlipped = 0
     val Flipped = 1
-    roadLink.attributes.getOrElse("MTKHEREFLIP", NotFlipped).asInstanceOf[Int] == Flipped
+    roadLink.attributes.getOrElse("MTKHEREFLIP", NotFlipped).asInstanceOf[BigInt] == Flipped
   }
 
   def generateValuesForLitRoads(): Unit = {
